requirements.txt
# Core Framework
Django==6.0
djangorestframework==3.15.2

# Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Phone Number Validation
django-phonenumber-field[phonenumbers]==8.0.0
phonenumbers==8.13.47

# CORS Headers
django-cors-headers==4.4.0

# Image Processing
Pillow==10.4.0

# Development Tools
django-debug-toolbar==4.4.6

# Python Standard Library Extensions
python-dateutil==2.9.0
pytz==2024.1
